vision-game/requirements/dev.txt
alembic==1.11.1
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
cffi==1.15.1
click==8.1.3
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.98.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
odtemplater==0.0.1
passlib==1.7.4
Pillow==9.5.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
